theme: Play with colors
Rather, remove them. I would like to try running with what is effectively no
font-locking for a while and see if there is a meaningful increase or decrease
in productivity.

I read somewhere that the color isn't too important; It's the consistency. I've
been meaning to try to see if it's possible to code with just italics and bold
alterations to the code.

The idea behind this is: We manipulate data (as variables), shaped like types.
So only the two should be highlighted. Everything else can be handled largely
via indentation. So try it out and see.