config: Make capture template insert a heading (defaulted to 2)
Dynamically sizing the heading is tricky, because
`--capture-template-interesting` is not called in the destination file, the one
which contains the markdown structure we want to parse.

We can fix this by leveraging the org capture template syntax `%(...)` but I'm
somehow getting `wholenump` errors during evaluation. I do not know yet what
caused that.

We can also leverage a `:before-finalize-hook`, but then the logic would be
split into two places, which might not be for the best.