config: transient: Move to prelude
Emacs has its own `transient` built-in, but it's out of date and not compatible
with the newer version.

There was a problem where any elisp files loaded before `config-git` that has a
`:after transient` will load the system transient. To prevent this, make
transient logically closer to utility libraries like dash and the like.

The only other package I've seen that behaves like this is `org`.