config: markdown, org-capture: Insert datetime headers in buffer
This avoids the trap of us inserting a datetree into a file, only to cancel the
capture, which does not remove the now-pointless datetree.

Now if we cancel captures it will also remove the relevant datetree headers.

There are _some_ issues with org capture template expansion, but I'm choosing to
not deal with that right now.