config: markdown: Update faces for metadata
Since git commits uses markdown, this means that the commit header is treated
like a metadata keypair (especially with the way I write it).

This means that this face is sort of fucked, since as a commit header I want it
to be a bit flashy, but for frontmatter metadata I want it to be as muted as
possible.

It's likely that there will be additional changes to this in the future.