config: markdown: transient: Save infix/suffix code that we don't use
I decided to move to a different strategy for not wanting to select the vault
directory all the time. One that is more compatible with org-capture and a bunch
of other things that coupling with transient will only complicate.

To be fair this whole attempt, while educational, is highlighting that I really
need to take a closer look at my markdown file access commands.